Fixing the database and data entry.
:wq
:wq
