+++ Explorate/__init__.py
@@ -1,8 +1,11 @@
 from flask import Flask
 from flask_migrate import Migrate
 from flask_sqlalchemy import SQLAlchemy
-from .models import db  # Import your models here
-from .routes import main  # Import your routes here
+from .models import db
+from .routes import main 
+
+
+# Create the db instance globally
 
 
 def create_app():
@@ -10,10 +13,13 @@ def create_app():
     app.config['SQLALCHEMY_DATABASE_URI'] = 'sqlite:///adventures.db'
     app.config['SQLALCHEMY_TRACK_MODIFICATIONS'] = False
 
+    # Initialize db with app
     db.init_app(app)
+
     migrate = Migrate(app, db)
 
     
     app.register_blueprint(main)
 
     return app
+
