+++ Explorate/static/js/script.js
