+++ Explorate/static/js/script_DataEnt.js
