+++ Explorate/static/js/share_script.js
