+++ create_db.py
@@ -1,10 +1,9 @@
-from app import create_app
-from app.models import db  # Import your models here
+from Explorate import create_app,db # Import the create_app function and db object
+from Explorate.models import db # Import all models
 
-# Create the app instance
+# Create the Flask app instance
 app = create_app()
 
-# Create all tables in the database (based on the models)
+# Create all tables in the database
 with app.app_context():
     db.create_all()
-    print("Database and tables created successfully!")
